moved class search modal to separate file
