+++ frontend/src/components/ClassSearchModal.js
@@ -1,0 +1,21 @@
+import React from 'react';
+import { Modal, Button } from 'react-bootstrap';
+
+function ClassSearchModal({ show, setShow }) {
+    const handleClose = () => setShow(false);
+    return (
+        <Modal show={show} onHide={handleClose} animation={false}>
+            <Modal.Header closeButton>
+                <Modal.Title>Modal heading</Modal.Title>
+            </Modal.Header>
+            <Modal.Body>Woohoo, you're reading this text in a modal!</Modal.Body>
+            <Modal.Footer>
+                <Button variant="secondary" onClick={handleClose}>
+                    Close
+                </Button>
+            </Modal.Footer>
+        </Modal>
+    );
+}
+
+export default ClassSearchModal;
